Introduced additional fonts for symbols and emoji.
Using some of the nerd fonts[1] urxvt can display glyphs, which are
available there, but still, there is a several other glyphs, which are
defined in different font files, than Deja Vu, which I'm using as a main
scalable font.

Moreover, I really like to use Fixed Misc[2] from Xorg, since it's crisp
and readable.

In this commit, urxvt will have a list of fonts passed to, instead of
only one, to have greater number of glyphs to display.

[1] https://www.nerdfonts.com/
[2] https://gitlab.freedesktop.org/xorg/font/font-misc-misc